refactor: move toolbar en contextMenu to new ui folder
