+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/editor.ts
@@ -1,7 +1,7 @@
 import { Editor } from '@tiptap/core'
 import { DEFAULT_EXTENSIONS, ExtensionType } from './extensions.ts'
-import { Toolbar } from './toolbar.ts'
-import { ContextMenu } from './contextMenu.ts'
+import { Toolbar } from './ui/toolbar.ts'
+import { ContextMenu } from './ui/contextMenu.ts'
 import { Modules, HtmlTransform, TiptapModule } from './types.ts'
 import { WysiwygOptions, WysiwygProfile } from '../wysiwyg/wysiwyg.ts'
 import { CkeditorStyle } from '../wysiwyg/ckeditorConfig.ts'
